moved getTextWidth/Height from prop obj to gc obj
